Move webpack config to root
